# Perf: optimize function cal_fvnl_dbeta