### Background
While profiling ABACUS with vtune and setting gamma_only to 1, I found that the `cal_fvnl_dbeta` function consumed a considerable amount of time (even more than cal_gint_force).
![image](https://github.com/deepmodeling/abacus-develop/assets/153698752/7f94158b-571c-436e-831e-c5181703a10a)
This is largely due to the overhead of vector allocations and deallocations within nested loops.
![img_v3_02cm_0858b00e-7eff-49e9-ac46-348a33aa240g](https://github.com/deepmodeling/abacus-develop/assets/153698752/d175faa9-ceea-48f4-944a-9804ce5656e8)
Optimizing this function by converting vectors to pointers dramatically decreased its runtime. Additionally, I replaced some `pow` function calls with direct multiplication calls because the `pow` function is rather slow. Through vtune, it became evident that when the exponent is a small integer, directly invoking multiplication operations can be dozens of times faster than calling pow.
Below is a table presenting the results of optimizing:
| | Before optimizing| After optimizing |
|----------|----------|----------|
| multiple-k | 15.39s | 8.48s |
| gamma_only | 94.92s | 55.73s |

the testing case is tests/performance/Si512, and the command is "OMP_NUM_THREADS=12 mpirun -n 1 abacus".